Frontend changes:    - Student portal results: dynamic per-component marks breakdown table with obtained/max per subject, fourth-subject italics, grade and pass/fail columns    - Exam routines: add considerForAnnualResult toggle in create/edit dialog, chip shown on routine card    - Audit logs: new AuditLogPage with date range filters, action type and submodule dropdowns, paginated list, expandable before/after JSON snapshots; wired into router     at /user-management/audit-logs    - Sidebar restructured: split Setup into Institute Setup (system settings, calendar, rooms, grading) and Academic Structure (academic years, shifts, classes,    sections, gender sections, groups, subjects, class routine); Rooms and Subjects moved out of Exams; Class Teachers moved to HR Management; Attendance section removed    from admin panel (teacher portal only)    - FBAC permission matrix: reorganised sections to mirror new sidebar structure; added 5 previously-missing submodules — ACADEMIC_CALENDAR, CLASS_ROUTINE,    CLASS_TEACHER, HR_TEACHER_DUTY (ATTENDANCE excluded as teacher-portal-only); improved submodule display labels
   Backend changes (academic_service):
   - Caffeine in-memory cache for student results: warm on result publish, evict on unpublish, 10-day TTL, keyed by enrollmentId:routineId
   - CacheWarmingService: async warm/evict scoped to active enrollments in the published routine's academic year
   - StudentRoutineResultResponse: added ComponentMark inner class with componentName, obtained, max; populated per subject in result query
   - AuditLog entity: added beforeValue and afterValue (LONGTEXT) columns for snapshot diffs
   - AuditLogService: async log() with 7-arg and 9-arg overloads delegating to private save() to avoid Spring AOP self-invocation issue
   - AuditHelper: static getUserId() via SecurityContextHolder, getIp() via RequestContextHolder (X-Forwarded-For aware), toJson() via Jackson
   - Audit calls added in: StudentServiceImpl (create/update/delete with before/after snapshots), StaffService (create/update/deactivate), SystemUserService
   (grant/suspend/reinstate), StudentMarkService (mark save), ExamRoutineServiceImpl (publish/unpublish results)
   - ExamRoutine: added consider_for_annual_result column (default true); annual result query filtered to only include routines with this flag set
   - Audit log search endpoint: paginated, filterable by date range, submodule, action type

   EOF
   )"

+++ src/main/java/com/example/academic_service/service/ClassTeacherService.java
@@ -26,14 +26,14 @@ public class ClassTeacherService {
     }
 
     public Map<String, Object> assign(Long staffId, Long sectionId, Integer academicYearId) {
-        if (classTeacherRepository.existsBySectionIdAndAcademicYearId(sectionId, academicYearId)) {
+        if (classTeacherRepository.existsBySectionIdAndAcademicYearId(sectionId.intValue(), academicYearId)) {
             throw new ResponseStatusException(HttpStatus.CONFLICT,
                     "This section already has a class teacher for this academic year");
         }
 
         Staff staff = staffRepository.findById(staffId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Staff not found"));
-        Section section = sectionRepository.findById(sectionId)
+        Section section = sectionRepository.findById(sectionId.intValue())
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Section not found"));
         AcademicYear year = academicYearRepository.findById(academicYearId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Academic year not found"));
